% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.R
\name{segment}
\alias{segment}
\title{Segment sentences into tokens}
\usage{
segment(sentence, model = NULL)
}
\arguments{
\item{sentence}{character vector.}

\item{model}{path to a model file if you have.}
}
\value{
a list.
}
\description{
Segment sentences into tokens
}
